Refactoring, flake8
