+++ .flake8
@@ -15,6 +15,7 @@ inline-quotes = "
 # Ignore the unused import in __init__.py (F401)
 # Asserts needed in test code (S101)
 per-file-ignores = __init__.py:F401
+                 = setup.py   :E501
                  = tests/*    :S101 
                  = test/*     :S101
